Refector: 자유게시판 > 글 작성 폼 코드 분리
