--- src/components/Board/BoardWrite/EditPost/EditPost.module.scss
@@ -1,83 +1,0 @@
-@use "@/styles/globals.scss" as *;
-
-.container {
-  padding-top: 40px;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  @include Size("mobile") {
-    padding-top: 16px;
-  }
-
-  @include Size("tablet") {
-    padding-top: 32px;
-  }
-
-  .center {
-    max-width: 880px;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-
-    .header {
-      width: 100%;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-
-      .title {
-        @include Title2;
-        @include Drag;
-
-        @include Size("mobile") {
-          @include Title3;
-        }
-      }
-    }
-
-    .categorys {
-      display: flex;
-      gap: 8px;
-      margin-top: 32px;
-      margin-bottom: 12px;
-
-      @include Size("mobile") {
-        margin-top: 20px;
-      }
-    }
-
-    .editor {
-      width: 100%;
-      min-height: 600px;
-      margin-top: 20px;
-
-      @include Size("mobile") {
-        min-height: 500px;
-      }
-
-      @include Size("tablet") {
-        min-height: 700px;
-      }
-    }
-
-    input[type="file"] {
-      margin-bottom: 10px;
-    }
-  }
-}
-
-.buttonContainer {
-  display: flex;
-  justify-content: center;
-  margin-top: 40px;
-  max-height: 200px;
-
-  .button {
-    width: 100%;
-    max-width: 200px;
-  }
-}
