+++ src/components/Board/BoardWrite/EditPost/EditPost.tsx
@@ -1,99 +1,79 @@
 import React, { useState, useRef, useEffect } from "react";
-import dynamic from "next/dynamic";
-import Script from "next/script";
 import { useRouter } from "next/router";
 
 import { useMutation } from "react-query";
 import { AxiosError } from "axios";
 
-import { postPresignedUrl } from "@/api/aws/postPresigned";
 import { usePostsDetails } from "@/api/posts/getPostsId";
 import { CreatePostRequest, putEditPosts } from "@/api/posts/putPostsId";
 
 import { useAuthStore } from "@/states/authStore";
 import { useModalStore } from "@/states/modalStore";
-import { useDeviceStore } from "@/states/deviceStore";
 
-import Button from "@/components/Button/Button";
-import TextField from "@/components/TextField/TextField";
 import Loader from "@/components/Layout/Loader/Loader";
+import PostForm from "@/components/Board/BoardWrite/PostForm/PostForm";
 
 import { useToast } from "@/hooks/useToast";
-import { useIsMobile } from "@/hooks/useIsMobile";
 
-import { imageUrl } from "@/constants/imageUrl";
-
-import type { EditPostProps } from "./EditPost.types";
-
-import styles from "./EditPost.module.scss";
-
-const Editor = dynamic(() => import("@tinymce/tinymce-react").then((mod) => mod.Editor), {
-  ssr: false,
-  loading: () => <Loader />,
-});
+interface EditPostProps {
+  id: string;
+}
 
 export default function EditPost({ id }: EditPostProps) {
   const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
   const user_id = useAuthStore((state) => state.user_id);
-  const isMobile = useDeviceStore((state) => state.isMobile);
-  const isTablet = useDeviceStore((state) => state.isTablet);
 
-  useIsMobile();
   const [title, setTitle] = useState("");
   const [content, setContent] = useState("");
   const [selectedCategory, setSelectedCategory] = useState("일반");
-  const [isScriptLoaded, setIsScriptLoaded] = useState(false);
   const { showToast } = useToast();
   const openModal = useModalStore((state) => state.openModal);
   const router = useRouter();
   const [hasUnsavedChanges, setHasUnsavedChanges] = useState(false);
-  const editorRef = useRef<any>(null);
   const hasUnsavedChangesRef = useRef(hasUnsavedChanges);
   const { data: posts, isLoading } = usePostsDetails(id as string);
 
-  const typeMap = {
+  const typeMapToKo = {
     NORMAL: "일반",
     QUESTION: "질문",
     FEEDBACK: "피드백",
   } as const;
 
   useEffect(() => {
-    if (window.tinymce) {
-      setIsScriptLoaded(true);
+    if (!isLoading) {
+      if (!isLoggedIn) {
+        showToast("로그인 후 작성 가능합니다.", "warning");
+        router.push("/");
+      } else if (user_id !== posts?.author.id) {
+        router.push("/");
+      }
     }
-  }, []);
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      showToast("로그인 후 작성 가능합니다.", "warning");
-      router.push("/");
-    } else if (user_id !== posts?.author.id) {
-      router.push("/");
-    }
-  }, [isLoggedIn, router, user_id, posts]);
+  }, [isLoggedIn, router, user_id, posts, isLoading, showToast]);
 
   useEffect(() => {
     if (posts) {
       setTitle(posts.title);
       setContent(posts.content);
-      setSelectedCategory(typeMap[posts.type as keyof typeof typeMap] || "일반");
+      setSelectedCategory(typeMapToKo[posts.type as keyof typeof typeMapToKo] || "일반");
     }
   }, [posts]);
 
   // 변경 사항 감지
   useEffect(() => {
-    const hasChanges =
-      title !== posts?.title || content !== posts?.content || selectedCategory !== posts?.type;
-    setHasUnsavedChanges(hasChanges);
+    if (posts) {
+      const originalType = typeMapToKo[posts.type as keyof typeof typeMapToKo] || "일반";
+      const hasChanges =
+        title !== posts.title || content !== posts.content || selectedCategory !== originalType;
+      setHasUnsavedChanges(hasChanges);
+    }
   }, [posts, title, content, selectedCategory]);
 
   // 브라우저 이벤트 핸들러
   useEffect(() => {
     const handleBeforeUnload = (e: BeforeUnloadEvent) => {
       if (hasUnsavedChanges) {
-        const message = "변경사항이 저장되지 않을 수 있습니다.";
         e.preventDefault();
-        return message;
+        return (e.returnValue = "변경사항이 저장되지 않을 수 있습니다.");
       }
     };
 
@@ -108,26 +88,26 @@ export default function EditPost({ id }: EditPostProps) {
 
   useEffect(() => {
     const handleRouteChangeStart = (url: string) => {
-      if (!hasUnsavedChangesRef.current) return;
+      if (hasUnsavedChangesRef.current) {
+        router.events.emit("routeChangeError");
 
-      router.events.emit("routeChangeError");
-
-      openModal({
-        type: null,
-        data: {
-          title: "수정을 취소하고 나가시겠어요?",
-          subtitle: "변경사항이 저장되지 않습니다",
-          confirmBtn: "나가기",
-          onClick: () => {
-            hasUnsavedChangesRef.current = false;
-            setHasUnsavedChanges(false);
-            router.push(url);
+        openModal({
+          type: null,
+          data: {
+            title: "수정을 취소하고 나가시겠어요?",
+            subtitle: "변경사항이 저장되지 않습니다",
+            confirmBtn: "나가기",
+            onClick: () => {
+              hasUnsavedChangesRef.current = false;
+              setHasUnsavedChanges(false);
+              router.push(url);
+            },
           },
-        },
-        isComfirm: true,
-      });
+          isComfirm: true,
+        });
 
-      throw "routeChange aborted.";
+        throw "routeChange aborted.";
+      }
     };
 
     router.events.on("routeChangeStart", handleRouteChangeStart);
@@ -167,7 +147,7 @@ export default function EditPost({ id }: EditPostProps) {
       return;
     }
 
-    const typeMap = {
+    const typeMapToEn = {
       일반: "NORMAL",
       질문: "QUESTION",
       피드백: "FEEDBACK",
@@ -177,7 +157,7 @@ export default function EditPost({ id }: EditPostProps) {
       await editPost({
         title,
         content,
-        type: typeMap[selectedCategory as keyof typeof typeMap],
+        type: typeMapToEn[selectedCategory as keyof typeof typeMapToEn],
       });
 
       hasUnsavedChangesRef.current = false;
@@ -185,174 +165,28 @@ export default function EditPost({ id }: EditPostProps) {
 
       router.push(`/posts/${id}`);
     } catch (error) {
-      if (error instanceof AxiosError) {
-        if (error.response?.status === 400) {
-          showToast("잘못된 요청입니다. 입력값을 확인해주세요.", "error");
-          return;
-        }
+      if (error instanceof AxiosError && error.response?.status === 400) {
+        showToast("잘못된 요청입니다. 입력값을 확인해주세요.", "error");
+        return;
       }
       showToast("수정 중 오류가 발생했습니다. 다시 시도해주세요.", "error");
     }
   };
 
-  const convertToWebP = async (file: File): Promise<File> => {
-    return new Promise((resolve, reject) => {
-      const img: HTMLImageElement = new window.Image();
-      img.src = URL.createObjectURL(file);
-      img.onload = () => {
-        const canvas = document.createElement("canvas");
-        canvas.width = img.width;
-        canvas.height = img.height;
-        const ctx = canvas.getContext("2d");
-        if (!ctx) return reject(new Error("Canvas context not available"));
-
-        ctx.drawImage(img, 0, 0);
-        canvas.toBlob((blob) => {
-          if (blob) {
-            const webpFile = new File([blob], file.name.replace(/\.[^.]+$/, ".webp"), {
-              type: "image/webp",
-            });
-            resolve(webpFile);
-          } else {
-            reject(new Error("Failed to convert image to WebP"));
-          }
-        }, "image/webp");
-      };
-      img.onerror = () => reject(new Error("Image loading failed"));
-    });
-  };
-
-  const disabled = !title.trim() || !content.trim() || isLoading;
-
   if (isLoading) return <Loader />;
 
   return (
-    <div className={styles.container}>
-      <Script
-        src="https://public.grimity.com/tinymce/tinymce.min.js"
-        onLoad={() => setIsScriptLoaded(true)}
-        strategy="afterInteractive"
-      />
-      <div className={styles.center}>
-        <section className={styles.header}>
-          <h2 className={styles.title}>글 수정하기</h2>
-        </section>
-        <section className={styles.categorys}>
-          {["일반", "질문", "피드백"].map((category) => (
-            <Button
-              key={category}
-              size="s"
-              type={selectedCategory === category ? "filled-primary" : "outlined-assistive"}
-              onClick={() => handleCategoryClick(category)}
-            >
-              {category}
-            </Button>
-          ))}
-        </section>
-        <TextField
-          placeholder="제목을 입력해주세요"
-          maxLength={32}
-          value={title}
-          onChange={handleTitleChange}
-        />
-        <section className={styles.editor}>
-          {isScriptLoaded ? (
-            <Editor
-              licenseKey="gpl"
-              onInit={(evt, editor) => (editorRef.current = editor)}
-              init={{
-                height: isMobile ? 500 : isTablet ? 700 : 600,
-                resize: "both",
-                menubar: false,
-                plugins: ["image", "link", "autolink"],
-                toolbar: isMobile
-                  ? "undo redo | h1 h2 link image | bold italic underline strikethrough | forecolor backcolor"
-                  : "undo redo | h1 h2 | bold italic underline strikethrough | forecolor backcolor | link image",
-                content_style: `
-                  body { 
-                    font-family: Pretendard, sans-serif; 
-                    font-size: 14px;                     
-                  }
-                  img { 
-                    max-width: 100%; 
-                    height: auto !important;
-                  }
-                  h1 {
-                    margin: 14px 0;
-                  }
-                  h2 {
-                    margin: 14px 0;
-                  }
-                  p {
-                    margin: 6px 0;
-                  }
-                `,
-                base_url: "https://public.grimity.com/tinymce",
-                skin_url: "https://public.grimity.com/tinymce/skins/ui/oxide",
-                icons_url: "https://public.grimity.com/tinymce/icons/default/icons.js",
-                statusbar: false,
-                indent: false,
-                indent_use_margin: true,
-                indent_size: 4,
-                setup: (editor) => {
-                  editor.on("keydown", (event) => {
-                    if (event.key === "Tab") {
-                      event.preventDefault();
-                      editor.execCommand("mceInsertContent", false, "&nbsp;&nbsp;&nbsp;&nbsp;");
-                    }
-                  });
-                },
-                images_upload_handler: async (
-                  blobInfo: { filename: () => string; blob: () => Blob },
-                  progress: (progress: number) => void,
-                ): Promise<string> => {
-                  try {
-                    const file = blobInfo.blob() as File;
-
-                    const ext = file.name.split(".").pop()?.toLowerCase() as "jpg" | "jpeg" | "png";
-                    const webpFile = await convertToWebP(file);
-
-                    const data = await postPresignedUrl({
-                      type: "post",
-                      ext: "webp",
-                    });
-
-                    const uploadResponse = await fetch(data.url, {
-                      method: "PUT",
-                      body: webpFile,
-                      headers: {
-                        "Content-Type": "image/webp",
-                      },
-                    });
-
-                    if (!uploadResponse.ok) {
-                      throw new Error(`${uploadResponse.status}`);
-                    }
-
-                    return `${imageUrl}/${data.imageName}`;
-                  } catch (error) {
-                    return Promise.reject("이미지 업로드 실패");
-                  }
-                },
-              }}
-              value={content}
-              onEditorChange={handleEditorChange}
-            />
-          ) : (
-            <Loader />
-          )}
-          <div className={styles.buttonContainer}>
-            <Button
-              className={styles.button}
-              type="filled-primary"
-              onClick={handleSubmit}
-              disabled={disabled}
-            >
-              수정 완료
-            </Button>
-          </div>
-        </section>
-      </div>
-    </div>
+    <PostForm
+      formTitle="글 수정하기"
+      title={title}
+      onTitleChange={handleTitleChange}
+      content={content}
+      onEditorChange={handleEditorChange}
+      selectedCategory={selectedCategory}
+      onCategoryClick={handleCategoryClick}
+      onSubmit={handleSubmit}
+      isSubmitting={isEditPostLoading}
+      submitButtonText="수정 완료"
+    />
   );
 }
