--- src/components/Board/BoardWrite/EditPost/EditPost.types.ts
@@ -1,3 +1,0 @@
-export interface EditPostProps {
-  id: string;
-}
