+++ src/components/Board/BoardWrite/PostForm/PostForm.module.scss
@@ -1,18 +1,17 @@
 @use "@/styles/globals.scss" as *;
 
 .container {
+  padding-top: 40px;
   width: 100%;
   height: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding-top: 60px;
-  padding-bottom: 80px;
-  padding-left: 32px;
-  padding-right: 32px;
 
   @include Size("mobile") {
-    padding: 16px;
+    padding-top: 16px;
+    padding-left: 16px;
+    padding-right: 16px;
   }
 
   @include Size("tablet") {
@@ -20,18 +19,25 @@
   }
 
   .center {
-    max-width: $max-width-content;
+    max-width: 880px;
     width: 100%;
     height: 100%;
     display: flex;
     flex-direction: column;
 
-    .title {
-      @include Title2;
-      @include Drag;
+    .header {
+      width: 100%;
+      display: flex;
+      align-items: center;
+      justify-content: space-between;
 
-      @include Size("mobile") {
-        @include Title3;
+      .title {
+        @include Title2;
+        @include Drag;
+
+        @include Size("mobile") {
+          @include Title3;
+        }
       }
     }
